# KYC-backend/api/cyopenapi.yaml
openapi: 3.0.0
info:
  title: CyServer API v3
  version: 3.1.0
  description: EBSI  v3
  termsOfService: https://api-pilot.ebsi.eu/docs/terms
  contact:
    name: EBSI Support - Service Desk
    url: https://ec.europa.eu/digital-building-blocks/tracker/plugins/servlet/desk/portal/11
  license:
    name: EUPL-1.2
    url: https://joinup.ec.europa.eu/page/eupl-text-11-12
servers:
  - url: http://testissuer.acgoldman.com:3005
  - url: http://localhost:3005
...
paths:
 /v3/auth/.well-known/openid-configuration:
    get:
      operationId: get-openid-configuration
      tags:
        - Auth Server
      summary: Authorisation Server discovery metadata
      description: |
        Client or a wallet must obtain the Authorisation Server metadata prior to the transaction using strategies described in [OAuth 2.0 Authorization Server Metadata](https://www.rfc-editor.org/info/rfc8414) or out-of-band.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/openid-configuration"
              examples:
                ebsiAuthoriseAndAccredit:
                  summary: |
                    Profile: EBSI Authorise and Accredit
                  value:
                    issuer: https://api-conformance.ebsi.eu/conformance/v3/auth
                    authorization_endpoint: https://api-conformance.ebsi.eu/conformance/v3/auth/authorize
                    token_endpoint: https://api-conformance.ebsi.eu/conformance/v3/auth/token
                    jwks_uri: https://api-conformance.ebsi.eu/conformance/v3/auth/jwks
                    redirect_uris:
                      - https://api-conformance.ebsi.eu/conformance/v3/auth/direct_post
                    scopes_supported:
                      - openid
                    response_types_supported:
                      - vp_token
                      - id_token
                    response_modes_supported:
                      - query
                    grant_types_supported:
                      - authorization_code
                    subject_types_supported:
                      - public
                    id_token_signing_alg_values_supported:
                      - ES256
                    request_object_signing_alg_values_supported:
                      - ES256
                    request_parameter_supported: true
                    request_uri_parameter_supported: true
                    token_endpoint_auth_methods_supported:
                      - private_key_jwt
                    request_authentication_methods_supported:
                      authorization_endpoint:
                        - request_object
                    vp_formats_supported:
                      jwt_vp:
                        alg_values_supported:
                          - ES256
                      jwt_vc:
                        alg_values_supported:
                          - ES256
                    subject_syntax_types_supported:
                      - did:key
                      - did:ebsi
                    subject_syntax_types_discriminations:
                      - did:key:jwk_jcs-pub
                      - did:ebsi:v1
                    subject_trust_frameworks_supported:
                      - ebsi
                    id_token_types_supported:
                      - subject_signed_id_token
                      - attester_signed_id_token
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/problem"
              example:
                type: about:blank
                title: Internal Server Error
                status: 500
                detail: |
                  Connection timeout
 /v3/auth/jwks:
    get:
      tags:
        - Auth Server
      summary: Authorisation Server's JWKS endpoint
      description: |
        A JSON Web Key (JWK) used by the authorisation server. [RFC4627].
      operationId: jwks
      responses:
        "200":
          description: Success
          content:
            application/jwk-set+json:
              schema:
                $ref: "#/components/schemas/jwks"
        "500":
          description: Internal Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/problem"
              examples:
                Internal Server Error:
                  value:
                    title: Internal Server Error
                    status: 500
                    detail: The server encountered an internal error and was unable to complete your request
 /v3/auth/authorize:
    get:
      operationId: get-authorize
      tags:
        - Auth Server
      summary: OIDC Authorization endpoint
      description: |
        After the discovery, the client proceeds with Verifiable Credential Issuance flow by requesting access for the required credential, from the Authorisation Server. The Authorisation Request Object must be signed with the client's private keys, owned by the requesting the client_id. The client's public key must be discoverable through client's openid-configuration through the jwks_uri parameter.
      parameters:
        - name: scope
          in: query
          description: |
            REQUIRED. OpenID Connect requests MUST contain the openid scope value. If the openid scope value is not present, the behavior is entirely unspecified.
          required: true
          schema:
            type: string
          examples:
            openid:
              summary: OpenID
              value: openid
        - name: response_type
          in: query
          required: true
          description: |
            REQUIRED. OAuth 2.0 Response Type value that determines the authorization processing flow to be used, including what parameters are returned from the endpoints used. When using the Authorization Code Flow, this value is code.

            MUST be 'code'
          schema:
            type: string
          example: code
        - name: client_id
          in: query
          required: true
          description: |
            OAuth 2.0 Client Identifier valid at the Authorization Server.
            
            For Holder wallet its DID
            
            For Verifiable Accreditation Issuance: MUST be URL of the issuer requesting the accreditation that was registered with the Accreditation Issuer
          schema:
            type: string
            
          examples:
            accreditationIssuance:
              summary: Accreditation or VC Issuance to Server/Cloud Wallet
              value: https://my-issuer.eu/suffix/xyz
        - name: redirect_uri
          in: query
          required: true
          description: |
            REQUIRED. Redirection URI to which the response will be sent. This URI MUST exactly match one of the Redirection URI values for the Client pre-registered at the OpenID Provider.
          schema:
            type: string
          example: https://my-issuer.eu/suffix/xyz/code-cb
        - name: state
          in: query
          description: |
            RECOMMENDED. Opaque value used to maintain state between the request
            and the callback. Typically, Cross-Site Request Forgery (CSRF, XSRF)
            mitigation is done by cryptographically binding the value of this
            parameter with a browser cookie.
          schema:
            type: string
          example: 32f6e80d-bc60-4261-a1c0-c6fe0b15bd9f
        - name: nonce
          in: query
          description: |
            OPTIONAL. String value used to associate a Client session with an ID Token, and to
            mitigate replay attacks. The value is passed through unmodified from the Authentication
            Request to the ID or VP Token. Sufficient entropy MUST be present in the nonce values
            used to prevent attackers from guessing values.
          schema:
            type: string
          example: n-0S6_WzA2Mj
        - name: request
          in: query
          required: false
          description: |
            Only for Service Wallets. Authorisation Request Object - The Request Object must be signed with the client's private keys, owned by the requesting client_id. The used private key's public key must be discoverable through client's openid-configuration through jwks_uri parameter.

            See the Authorisation Request Object schema.
          schema:
            type: string
            format: '^(([A-Za-z0-9\-_])+\.)([A-Za-z0-9\-_]+)(\.([A-Za-z0-9\-_]+))?$'
          example: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6Im1jX0pXbHc3ZTZjUnM5NFhTUXMxWGEyeWh3U1FQMm1XellCZ1NRY2lSOTAifQ.eyJzY29wZSI6Im9wZW5pZCIsImNsaWVudF9pZCI6Imh0dHBzOi8vY29uZm9ybWFuY2UtdGVzdC5lYnNpLmV1L2NvbmZvcm1hbmNlL3YzL2NsaWVudC1tb2NrL2RpZDplYnNpOnpnQU1Ya1oyZk00SnNSU2l4alY2QkpoIiwicmVkaXJlY3RfdXJpIjoiaHR0cHM6Ly9jb25mb3JtYW5jZS10ZXN0LmVic2kuZXUvY29uZm9ybWFuY2UvdjMvY2xpZW50LW1vY2svZGlkOmVic2k6emdBTVhrWjJmTTRKc1JTaXhqVjZCSmgvY29kZS1jYiIsInJlc3BvbnNlX3R5cGUiOiJjb2RlIiwic3RhdGUiOiI0MGQzYzZmNS0wMTY0LTQ3OTctOTg5ZS1kNWY4ZTg2YTAxMmYiLCJjbGllbnRfbWV0YWRhdGEiOnsiandrc191cmkiOiJodHRwczovL2NvbmZvcm1hbmNlLXRlc3QuZWJzaS5ldS9jb25mb3JtYW5jZS92My9jbGllbnQtbW9jay9kaWQ6ZWJzaTp6Z0FNWGtaMmZNNEpzUlNpeGpWNkJKaC9qd2tzIiwiYXV0aG9yaXphdGlvbl9lbmRwb2ludCI6Imh0dHBzOi8vY29uZm9ybWFuY2UtdGVzdC5lYnNpLmV1L2NvbmZvcm1hbmNlL3YzL2NsaWVudC1tb2NrL2RpZDplYnNpOnpnQU1Ya1oyZk00SnNSU2l4alY2QkpoL2F1dGhvcml6ZSJ9LCJhdXRob3JpemF0aW9uX2RldGFpbHMiOlt7InR5cGUiOiJvcGVuaWRfY3JlZGVudGlhbCIsImZvcm1hdCI6Imp3dF92YyIsImxvY2F0aW9ucyI6WyJodHRwczovL2NvbmZvcm1hbmNlLXRlc3QuZWJzaS5ldS9jb25mb3JtYW5jZS92My9pc3N1ZXItbW9jayJdLCJ0eXBlcyI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsIlZlcmlmaWFibGVBdHRlc3RhdGlvbiIsIlZlcmlmaWFibGVBdXRob3Jpc2F0aW9uVG9PbmJvYXJkIl19XSwiYXVkIjoiaHR0cHM6Ly9jb25mb3JtYW5jZS10ZXN0LmVic2kuZXUvY29uZm9ybWFuY2UvdjMvYXV0aC1tb2NrIiwiaXNzIjoiaHR0cHM6Ly9jb25mb3JtYW5jZS10ZXN0LmVic2kuZXUvY29uZm9ybWFuY2UvdjMvY2xpZW50LW1vY2svZGlkOmVic2k6emdBTVhrWjJmTTRKc1JTaXhqVjZCSmgifQ.bdBxZTbo5_THLdcLndNVf3ZGSBDTKqGnyK-E14Ah94lqjIk16DjSGNvibSaoGFeNGGknXttEvoZUSiEMh5ke7A
        - name: authorization_details
          in: query
          required: false
          description: |
            Only for Holder Wallets.  OID authorisation details data model.

            Note: `authorization_details` must be a stringified JSON array.

            See "OID4VCI Authorisation Details" schema for more information.
          schema:
            type: string
          example: '[{"type":"openid_credential","format":"jwt_vc","locations":["http://localhost:3000/conformance/v3/issuer"],"types":["VerifiableCredential","VerifiableAttestation","CTWalletCrossInTime"]}]'
        - name: client_metadata
          in: query
          required: false
          description: |
            Only for Holder Wallets. Client Metadata including a link to an `authorization_endpoint` (optional).

            Note: `client_metadata` must a stringified JSON object.

            See "Client Metadata" schema for more information.
          schema:
            type: string
          example: '{"authorization_endpoint":"openid:"}'
        - name: issuer_state
          in: query
          required: false
          description: |
            REQUIRED if Credential Offering contained `issuer_state`.
          schema:
            type: string
          example: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6Ink2UTFUTWN4am4zU1ZDT0dNUm9mZHA0M1owU0tGMVROSTkwcG5CWldSZkkifQ.eyJjbGllbnRfaWQiOiJkaWQ6a2V5OnoyZG16RDgxY2dQeDhWa2k3SmJ1dU1tRllyV1BnWW95dHlrVVozZXlxaHQxajlLYnB0SnZwaHdHaXh6azJiM1l4ZnNxRzZwRFNYTFhia0g1Y01HYjhBZDFWSkp5ZWkxWWhEQjRDb2RNNW5DaDJxUGU4TFhNQVU2OWdjdTZmUmI2TExvYmQxRmZjOXdSQlU0NXBIWWNKUUVnamg4blZVdmR3NUhqWGgzNEJ3NUpQNzR0N1ciLCJjcmVkZW50aWFsX3R5cGVzIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsIiwiVmVyaWZpYWJsZUF0dGVzdGF0aW9uIiwiQ1RXYWxsZXRDcm9zc0luVGltZSJdLCJpYXQiOjE2ODI0OTY3MTIsImV4cCI6MTY4MjQ5NzAxMiwiaXNzIjoiaHR0cDovL2xvY2FsaG9zdDozMDAwL2NvbmZvcm1hbmNlL3YzL2lzc3Vlci1tb2NrIiwiYXVkIjoiaHR0cDovL2xvY2FsaG9zdDozMDAwL2NvbmZvcm1hbmNlL3YzL2F1dGgtbW9jayIsInN1YiI6ImRpZDprZXk6ejJkbXpEODFjZ1B4OFZraTdKYnV1TW1GWXJXUGdZb3l0eWtVWjNleXFodDFqOUticHRKdnBod0dpeHprMmIzWXhmc3FHNnBEU1hMWGJrSDVjTUdiOEFkMVZKSnllaTFZaERCNENvZE01bkNoMnFQZThMWE1BVTY5Z2N1NmZSYjZMTG9iZDFGZmM5d1JCVTQ1cEhZY0pRRWdqaDhuVlV2ZHc1SGpYaDM0Qnc1SlA3NHQ3VyJ9.QzNioxnUbdxhdnYoAzNiati7-Lg5Kkg0mEAf1OHRvXyjxn1Z-_kuvcVdOmIpsF_o55l4NMy2hmz5yLIFs9zVEA
        - name: code_challenge
          in: query
          required: false
          description: |
            Only for Holder Wallets. In format of `BASE64URL-ENCODE(SHA256(code_verifier as UTF-8 string))`.

            `code_verifier` is client generated secure random, which will be used with token endpoint. It is between 43 and 128 characters long, and contains characters A-Z, a-z, 0-9, hyphen, period, underscore, and tilde. Please see [RFC 7636](https://datatracker.ietf.org/doc/html/rfc7636#section-4.1)
          schema:
            type: string
          example: yKZeDTMFPu8YSwqdlIVkQBifcwmtUgnqzjP-hvV0qdI
        - name: code_challenge_method
          in: query
          required: false
          description: |
            Only for Holder Wallets. MUST be "S256".
          schema:
            type: string
          example: S256
      responses:
        "302":
          description: |
            Authorisation Server responds with one of the three responses:
              - ID Token Request
              - VP Token Request
              - Error codes for authorization endpoint

            All responses are in the "Location" header parameter and are x-www-form-urlencoded
          headers:
            Location:
              description: x-www-form-urlencoded ID Token Request
              required: true
              schema:
                type: string
                format: URL
              examples:
                success:
                  summary: Successful Request
                  value: openid://?state=32f6e80d-bc60-4261-a1c0-c6fe0b15bd9f&client_id=https%3A%2F%2Fconformance-test.ebsi.eu%2Fconformance%2Fv3%2Fauth&redirect_uri=https%3A%2F%2Fconformance-test.ebsi.eu%2Fconformance%2Fv3%2Fauth%2Fdirect_post&response_type=id_token&response_mode=direct_post&scope=openid&nonce=8b4b9a89-2d0d-4b94-b975-3c93af758926&request=eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6IkZMeEkzTE04bUZDRkNEMUg0VmpacVd0MVBmaWQyaThBQ1lpRHZFelo5VU0ifQ.eyJzdGF0ZSI6IjMyZjZlODBkLWJjNjAtNDI2MS1hMWMwLWM2ZmUwYjE1YmQ5ZiIsImNsaWVudF9pZCI6Imh0dHBzOi8vY29uZm9ybWFuY2UtdGVzdC5lYnNpLmV1L2NvbmZvcm1hbmNlL3YzL2F1dGgtbW9jayIsInJlZGlyZWN0X3VyaSI6Imh0dHBzOi8vY29uZm9ybWFuY2UtdGVzdC5lYnNpLmV1L2NvbmZvcm1hbmNlL3YzL2F1dGgtbW9jay9kaXJlY3RfcG9zdCIsInJlc3BvbnNlX3R5cGUiOiJpZF90b2tlbiIsInJlc3BvbnNlX21vZGUiOiJkaXJlY3RfcG9zdCIsInNjb3BlIjoib3BlbmlkIiwibm9uY2UiOiI4YjRiOWE4OS0yZDBkLTRiOTQtYjk3NS0zYzkzYWY3NTg5MjYiLCJpc3MiOiJodHRwczovL2NvbmZvcm1hbmNlLXRlc3QuZWJzaS5ldS9jb25mb3JtYW5jZS92My9hdXRoLW1vY2siLCJhdWQiOiJodHRwczovL215LWlzc3Vlci5ldS9zdWZmaXgveHl6In0.WwQFrOulUdXt2Yr-rqD-8xHCBd1EZRVM6KaVWWsZJGWh7mKmj6qDJrlGjWsJUcLYgF9AaetjOxubKeeQ4VA60Q
                invalidRequest:
                  summary: Invalid Request
                  value: https://my-issuer.eu/suffix/xyz/code-cb?error=invalid_request&error_description=%5Binvalid_literal%5D+in+%27request.authorization_details.0.type%27%3A+Invalid+literal+value%2C+expected+%22openid_credential%22%0A%5Binvalid_literal%5D+in+%27request.authorization_details.0.format%27%3A+Invalid+literal+value%2C+expected+%22jwt_vc%22%0A%5Bcustom%5D+in+%27request.authorization_details.0.types%27%3A+Array+must+include+VerifiableCredential%2C+VerifiableAttestation&state=32f6e80d-bc60-4261-a1c0-c6fe0b15bd9f
        "400":
          description: |
            Authorisation Endpoint Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/problem"
              examples:
                invalidRedirectURI:
                  value:
                    detail: '["redirect_uri must be a URL address"]'
                    status: 400
                    title: "Bad Request"
                    type: about:blank
        "500":
          description: Internal Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/problem"
              examples:
                Internal Server Error:
                  value:
                    title: Internal Server Error
                    status: 500
                    detail: The server encountered an internal error and was unable to complete your request
 /v3/auth/request_uri/{requestId}:
    get:
      operationId: get-request_uri
      tags:
        - Auth Server
      summary: Request by reference
      description: |
        Authorisation Server's endpoint that returns a [Request Object](https://openid.net/specs/openid-connect-core-1_0.html#UseRequestUri).
        request_uri SHOULD be used whenever the request object is large.
      parameters:
        - name: requestId
          in: path
          description: Unique Request ID
          required: true
          schema:
            type: string
          examples:
            tokenRequestID:
              summary: Request ID
              value: 2160d1e4-2243-48cc-ae90-169e3d206502
      responses:
        "200":
          description: OK
          content:
            application/jwt:
              schema:
                title: Request
                description: |
                  Authorisation Request Object - The Request Object must be signed with the client's (Credential Issuer) private keys, owned by the requesting client_id. The used private key's public key must be discoverable through client's openid-configuration through jwks_uri parameter.

                  The Authorisation Request Object MUST be a valid JWS.

                  See the Authorisation Request Object schema.
                type: string
                format: '^(([A-Za-z0-9\-_])+\.)([A-Za-z0-9\-_]+)(\.([A-Za-z0-9\-_]+))?$'
                example: |
                  eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6IkZMeEkzTE04bUZDRkNEMUg0VmpacVd0MVBmaWQyaThBQ1lpRHZFelo5VU0ifQ.eyJzdGF0ZSI6IjIyYmQxM2FiLTM1NjYtNDM1MS04MDBiLWExNGM1Y2Q3ZjE3NiIsImNsaWVudF9pZCI6Imh0dHBzOi8vY29uZm9ybWFuY2UtdGVzdC5lYnNpLmV1L2NvbmZvcm1hbmNlL3YzL2F1dGgtbW9jayIsInJlZGlyZWN0X3VyaSI6Imh0dHBzOi8vY29uZm9ybWFuY2UtdGVzdC5lYnNpLmV1L2NvbmZvcm1hbmNlL3YzL2F1dGgtbW9jay9kaXJlY3RfcG9zdCIsInJlc3BvbnNlX3R5cGUiOiJpZF90b2tlbiIsInJlc3BvbnNlX21vZGUiOiJkaXJlY3RfcG9zdCIsInNjb3BlIjoib3BlbmlkIiwibm9uY2UiOiIyMjgwMGEyMS0zNGNhLTQ5Y2MtYTNhMS00ODRjNzdhNzhiZTAiLCJpc3MiOiJodHRwczovL2NvbmZvcm1hbmNlLXRlc3QuZWJzaS5ldS9jb25mb3JtYW5jZS92My9hdXRoLW1vY2siLCJhdWQiOiJodHRwczovL215LWlzc3Vlci5ldS9zdWZmaXgveHl6In0.9cWdEYiZNGLO0pmSwnYCvMfECGkV7egVRqG4LMgBO7y1wKVz23QmrPOR_QbpxVA-IjkbSqWdXeTuVsNSSGlAlw
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/problem"
              examples:
                invalidState:
                  value:
                    detail: '["requestId must be a UUID"]'
                    status: 400
                    title: Bad Request
                    type: about:blank
        "404":
          description: Request Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/problem"
              examples:
                invalidState:
                  value:
                    detail: No Authorization Request found with the ID {requestId}
                    status: 404
                    title: Not Found
                    type: about:blank
        "500":
          description: Internal Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/problem"
              examples:
                Internal Server Error:
                  value:
                    title: Internal Server Error
                    status: 500
                    detail: The server encountered an internal error and was unable to complete your request
 /v3/auth/direct_post:
    post:
      operationId: post-direct_post
      tags:
        - Auth Server
      summary: Direct Post
      description: |
        Authorisation Server's endpoint that accepts and processes the ID Token issued by the client. The ID Token is self-issued by the client and it MUST be signed with client's DID document's authentication key.

        See the ID Token - Direct POST response schema.
      requestBody:
        content:
          x-www-form-urlencoded:
            schema:
              type: object
              description: |
                Self-issued ID Token and signed with DID Authentication key or VP Token. The signed ID Token or VP Token is submitted as compact serialised JWS.
              required:
                - state
              properties:
                id_token:
                  description: ID Token. Required when the Authorisation Server requests an ID Token.
                  type: string
                  format: '^(([A-Za-z0-9\-_])+\.)([A-Za-z0-9\-_]+)(\.([A-Za-z0-9\-_]+))?$'
                  example: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6ImRpZDplYnNpOnpybk1TNkJXZHB6UTdtelk0RXdmVFZmIn0.eyJub25jZSI6IjBhMjMwM2VhLTlkNTctNDRiMC04NjU3LTAyYjBkMDkyZjg3MyIsInN1YiI6Imh0dHBzOi8vbXktaXNzdWVyLmV1L3N1ZmZpeC94eXoiLCJpc3MiOiJodHRwczovL215LWlzc3Vlci5ldS9zdWZmaXgveHl6IiwiYXVkIjoiaHR0cHM6Ly9jb25mb3JtYW5jZS10ZXN0LmVic2kuZXUvY29uZm9ybWFuY2UvdjMvYXV0aC1tb2NrIiwiaWF0IjoxNjc2NjQxNzE5LCJleHAiOjE2NzY2NDIwMTl9.N3tAGPBZ-hFkzF-c8Gr6C-YVgvrjbfwqU4iVnQJtCPcq4qioUnr-I9flwubJJYHfbqkC1MoaYGOq5X69kL69nQ
                vp_token:
                  description: VP Token. Required when the Authorisation Server requests an VP Token.
                  type: string
                  format: '^(([A-Za-z0-9\-_])+\.)([A-Za-z0-9\-_]+)(\.([A-Za-z0-9\-_]+))?$'
                presentation_submission:
                  description: Stringified presentation submission object. Required when the Authorisation Server requests an VP Token.
                  type: string
                state:
                  type: string
                  example: 8379a190-e437-49f8-8585-a9aa55815c5
      responses:
        "302":
          description: OK
          headers:
            Location:
              description: |
                x-www-form-urlencoded Authentication Response. It MUST contain a valid `code` and `state`. The `state` claim MUST have the same value as the `state` in the Authorisation Request.

                See Authentication Response schema definition.
              schema:
                type: string
        "400":
          description: |
            Authorisation Endpoint Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/problem"
              examples:
                invalidState:
                  value:
                    detail: No Authentication Request bound to state ${state} found
                    status: 400
                    title: Bad Request
                    type: about:blank
        "500":
          description: Internal Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/problem"
              examples:
                Internal Server Error:
                  value:
                    title: Internal Server Error
                    status: 500
                    detail: The server encountered an internal error and was unable to complete your request
 /v3/auth/token:
    post:
      operationId: post-token
      tags:
        - Auth Server
      summary: Token endpoint
      description: |
        The client proceeds with the code flow, and calls the Token Endpoint with the required details and signs client_assertion JWT with client's private keys, which public key counterparts are resolvable through jwks_uri or is in the Client Metadata shared in the pre-registration step.
      requestBody:
        content:
          x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/token_request"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/access_token_response"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/problem"
              examples:
                Bad Request:
                  value:
                    title: Bad Request
                    status: 400
                    detail: Bad request.
                Token Expired:
                  value:
                    title: Token Expired
                    status: 400
                    detail: The token has expired.
                Issuer Not Found:
                  value:
                    title: Issuer Not Found
                    status: 400
                    detail: Issuer not found in the trusted apps registry.
        "401":
          description: Unauthorised
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/problem"
              examples:
                Unauthorised:
                  value:
                    title: Unauthorized
                    status: 401
                    detail: You are not authorized to access the resources.
        "500":
          $ref: "#/paths/~1v3~1auth~1jwks/get/responses/500"
...
components:
  schemas:
...
    problem:
      title: Problem Details
      description: A Problem Details object (RFC 7807)
      type: object
      properties:
        type:
          type: string
          format: uri
          description: |
            An absolute URI that identifies the problem type. When dereferenced, it SHOULD provide
            human-readable documentation for the problem type.
          default: about:blank
        title:
          type: string
          description: A short summary of the problem type.
          example: Internal Server Error
        status:
          type: integer
          format: int32
          description: |
            The HTTP status code generated by the origin server for this occurrence of the problem.
          minimum: 400
          maximum: 600
          example: 500
        detail:
          type: string
          description: |
            A human readable explanation specific to this occurrence of the problem.
          example: Connection timeout
        instance:
          type: string
          format: uri
          description: |
            An absolute URI that identifies the specific occurrence of the problem. It may or may
            not yield further information if dereferenced.
...
    jwks:
      title: JSON Web Key Set
      required:
        - keys
      type: object
      description: |
        JSON Web Key Set (JWKS) - A JSON object that represents a set of JWKs. The JSON object MUST have a keys member, which is an array of JWKs.
      properties:
        keys:
          type: array
          description: |
            List of JSON Web Key (JWK) - A JSON object that represents a cryptographic key. The members of the object represent properties of the key, including its value.
          items:
            title: JSON Web Key
            required:
              - kty
            type: object
            properties:
              kty:
                type: string
              crv:
                type: string
              alg:
                type: string
              n:
                type: string
              e:
                type: string
              x:
                type: string
              y:
                type: string
              kid:
                type: string
      example:
        keys:
          - kty: EC
            crv: P-256
            alg: ES256
            x: FMWOWBx2VroVMZVmn-TOU_943V7eYKQDl3GpDFjLQd4
            y: 7EGGmZW9ZhhJQUxSPH3_mpbxyNvC5eeBchGPkccZmGk
            kid: FLxI3LM8mFCFCD1H4VjZqWt1Pfid2i8ACYiDvEzZ9UM
...
    access_token_response:
      title: Access Token Response
      type: object
      description: Access Token Response
      required:
        - access_token
        - id_token
        - token_type
        - expires_in
        - c_nonce
        - c_nonce_expires_in
      properties:
        access_token:
          title: Access Token
          description: |
            The access token issued by the authorization server.

            The Access Token MUST be signed by the authorisation server.
          type: string
          format: JWS
        id_token:
          title: ID Token
          description: |
            The ID token issued by the authorization server.
          type: string
          format: JWS
        token_type:
          title: Token Type
          description: |
            The access token type provides the client with the information required to successfully utilize the access token to make a protected resource request (along with type-specific attributes).
          type: string
          enum:
            - bearer
        expires_in:
          title: Expires in
          type: integer
          description: |
            The lifetime in seconds of the access token. For example, the value \"3600\" denotes that the access token will expire in one hour from the time the response was generated.

            Lifetime must be 86400 seconds.
        c_nonce:
          title: Challenge Nonce
          description: |
            JSON string containing a nonce to be used to create a proof of possession of key material when requesting a Credential (see Section 7.2 of OpenID for VCI). When received, the Wallet MUST use this nonce value for its subsequent credential requests until the Credential Issuer provides a fresh nonce.
          type: string
        c_nonce_expires_in:
          title: Challenge Nonce expires in
          type: integer
          description: |
            The lifetime in seconds of the c_nonce token. For example, the value \"3600\" denotes that the access token will expire in one hour from the time the response was generated.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT